feat: Split VariableResolver behaviour to 2 classes
