+++ src/init/variables/computers/ArgExtractor.ts
@@ -1,9 +1,9 @@
-import { VarComputer } from '../VarComputer';
+import { ValueComputer } from './ValueComputer';
 
 /**
  * Simple VarComputer that just extracts an arg from parsed args.
  */
-export class ArgExtractor extends VarComputer {
+export class ArgExtractor extends ValueComputer {
   private readonly argKey: string;
 
   public constructor(argKey: string) {
