+++ src/init/variables/computers/ValueComputer.ts
@@ -1,0 +1,6 @@
+import { AsyncHandler } from '../../../util/handlers/AsyncHandler';
+
+/**
+ A handler that determines the value of a specific variable from parsed cli arguments.
+ */
+export abstract class ValueComputer extends AsyncHandler<Record<string, unknown>, unknown> {}
